restructured endpoints, refactored code, added worker hostname
