+++ src/constants/constants.py
@@ -1,5 +1,6 @@
 LOGS_DIR = "logs"
-ID_REGEX = "(?P<id_>[a-f0-9]+)"
+HEX_REGEX = "(?P<{}>[a-f0-9]+)"
+STRING_REGEX = "(?P<{}>[^()]+)"
 TIMESTAMP_FORMAT = "%Y-%m-%d %H:%M:%S"
 
 # application specific constants
