+++ src/constants/db_keys.py
@@ -17,3 +17,4 @@ SUCCESS = "success"
 GRADING_RUN = "grading_run_id"
 STAGES = "stages"
 JOB_ID = "job_id"
+WORKER_HOSTNAME = "worker_hostname"
